fix: cast models to weight_dtype in build_gradient_map
Models are loaded as float16 in main() and only cast to weight_dtype
(bfloat16) inside pgd_attack(). build_gradient_map() runs before
pgd_attack(), so model weights were float16 while input tensors were
bfloat16 — causing the matmul dtype mismatch.

Fix: add dtype=weight_dtype to all .to() calls for models inside
build_gradient_map.

+++ attacks/mist.py
@@ -650,12 +650,12 @@ def build_gradient_map(
     try:
         is_sdxl = text_encoder_2 is not None and tokenizer_2 is not None
 
-        # Ensure all models are on the target device before any forward pass
-        vae.to(device)
-        text_encoder.to(device)
-        unet.to(device)
+        # Ensure all models are on the target device and dtype before any forward pass
+        vae.to(device, dtype=weight_dtype)
+        text_encoder.to(device, dtype=weight_dtype)
+        unet.to(device, dtype=weight_dtype)
         if is_sdxl:
-            text_encoder_2.to(device)
+            text_encoder_2.to(device, dtype=weight_dtype)
 
         # 1. Load and resize instance image
         img = Image.open(instance_image_path).convert("RGB").resize((resolution, resolution))
